Break tests into smaller pieces
