+++ Statsdash/tests/test_date_utils.py
@@ -1,0 +1,10 @@
+from datetime import datetime, date, timedelta
+import unittest
+
+from Statsdash.utils import date as date_utils
+
+
+class TestDateUtils(unittest.TestCase):
+
+    def setUp(self):
+        pass
